+++ CHANGES.rst
@@ -44,6 +44,7 @@ Development
 - Ship configuration blueprint with deep sleep disabled
 - Improve MiniNet helper to get IP address
 - Set default maintenance duty cycle to 15 seconds
+- Improve Watchdog subsystem by adjusting to edge cases
 
 
 2019-06-22 0.5.1